Add a tag/category system.
The first thing I tried was merging this PR:
https://github.com/barryclark/jekyll-now/pulls/186
using the following procedure:

    curl https://github.com/barryclark/jekyll-now/commit/8615467a04f44b0f336069feb6c573a37c76590b.patch \
        | patch -p 1
    git add _sass/_tags.scss

Unfortunately, it didn't look very good. So I ripped it back out
and what I ended up using was the tagging system described in this
blog post by Zhihao Li:
http://codinfox.github.io/dev/2015/03/06/use-tags-and-categories-in-your-jekyll-based-github-pages/
The source code is available at
https://github.com/codinfox/codinfox-lanyon

I made a few modifications to the algorithmic level for clarity and speed,
but basically I'm using Li's code, especially the use of FontAwesome for
the tag icon itself.

This commit also contains a lot of security-minded cleanup. Some of
the cleanup was actually necessary in order to render tags with names
like `c++` and post titles containing Markdown (e.g. code keywords);
and some of the cleanup was just me being detail-oriented.