Patterns.
Thanks to this StackOverflow answer for explaining how to resize the iframe:
https://stackoverflow.com/a/701521/1424877

If you're reading this, you might be interested to know that you
can "draw" onto the canvas with your mouse and then run the CA on
the shapes you just drew. At least if you have a mouse. I don't think
mobile users get to play with that particular toy, unfortunately.